upf: in PFCP Sess Est, use local F-TEID = CHOOSE
So far the test asked osmo-upf to use an actual IP address and TEID, but
normally it is sent with CHOOSE = 1 so that osmo-upf picks a new local
TEID and returns it with the proper local IP address in the response.

This is making the test more realistic, and also prepares for testing
Network Instance handling, where osmo-upf will need to return the proper
IP addresses for specific Network Instance names.

Change-Id: Iaba372cbb2e0cf0c5bee80b09d346f9bcb78bfbe