bts: Proper tear down
This way we stop vc_IPA_OML before stopping the main component.
If we don't do this, the vc_IPA_OML is potentially receiving tons of OML
messages during BTS bringup which it tries to forward to the main
component. If the main component has stopped at that point, port
IPA_OML_PORT is no longer active and hence vc_IPA_OML fails with a
dynamic test error.

Change-Id: I761ae94aa698a74bb30b8a6b1860358545f222b5