ggsn: Speed up tests by lowering retransmission T3,N3 timers
The params are set in GGSN_Tests.cfg and not in GGSN_Tests.default,
because those lower values are not really good default ones as per spec.
For instance when running against open5gs-smfd, these values should be
left to the default in code of 5 and 3 seconds, because open5gs-smfd
doesn't support changing the values through cfg file yet.

Related: OS#5485
Change-Id: I3798fba89c2c357afeaa83a73759442c6c433cea