hnbgw: wait longer for PFCP Assoc
Since I introduced stricter timeouts on PFCP messages in
    hnbgw: add f_pfcp_expect()
    commit 6bbfe057afe6f23e2f325789a3dcac81065ab952
    Change-Id I2117475b695d486b1204d61e5bb21120a6187354
the HNBGW tests with PFCP support fail often:

The Assoc Setup resending timeout (X26) is configured to 5 seconds, and
the timeout to wait for this is also 5 seconds. Every once in a while
they happen to miss, causing a test failure.

Increase the initial Assoc Setup Req timeout to 15 seconds to avoid
sporadic failures.

Change-Id: I4b9af224e2346a4735f3d4e01b234acf56c5f3ff