hnbgw: give RAN startup some time
We have sporadic test failure, because the test starts to send RUA
messages before the CN link was able to perform a RANAP RESET + ACK
procedure.

Since recently, osmo-hnbgw is stricter on RANAP RESET: it will only
start passing RUA connections to the CN when the CN link has seen a
RANAP RESET ACK (from either side). So now, when the test is too quick,
its first RUA message is dropped on the floor, and the test fails, since
the RANAP never turns up on SCCP.

This is a quick hack, better would be to wait for some signal from the
RAN_Emulation when we are ready.

This jenkins trace shows that the RUA InitialUE happens about 200 ms
before RESET ACK is complete:

 Time                Protocol Info
 04:02:17.088852     RANAP    Reset
 04:02:17.105820     RANAP    Reset
 04:02:17.899887     RANAP    (RUA) InitialUE-Message (DTAP) (Unknown)
 04:02:17.905122     RANAP    Reset
 04:02:17.906043     RANAP    SACK (Ack=2, Arwnd=106496) ResetAcknowledge
 04:02:17.906056     RANAP    SACK (Ack=5, Arwnd=106436) ResetAcknowledge
 04:02:18.081522     RANAP    Reset ResetAcknowledge
 04:02:18.082829     RANAP    SACK (Ack=4, Arwnd=106496) ResetAcknowledge

https://jenkins.osmocom.org/jenkins/job/ttcn3-hnbgw-test/512/artifact/logs/hnbgw-tester/HNBGW_Tests.TC_rab_assign_mgcp_to.pcap.gz

Change-Id: Icbe7220112fbfe4ff5a5e1b9b65eeec428e51530