SGSN_Tests: add RIM ROUTING ADDRESS field in request
When sending the RAN INFORMATION REQUEST on GTP-C, we do not add a RIM
ROUTING ADDRESS + RIM ROUTING ADDRESS DISCRIMINATOR field. Those fields
are optional but commonly added so that the SGSN does not have to
disassemble the RAN TRANSPARENT CONTAINTER to know the destination
address.

see also: 3gpp TS 29.060, section 7.5.14 and
          3GPP TS 23.060, section 8.1.5.2.2

Change-Id: Id944c66f28d787a18c6c6f7c9dc885997d83e94c
Related: OS#6095