NAS_Templates: fix ts_NAS_GUTI
The template ts_NAS_GUTI permutates the MCC/MNC digits in a weird way,
which seems to map to a format that is not used anywhere else. Also the
template is not used anywhere yet.

Let's not permutate the MCC/MNC digit, instead let's put a comment that
makes clear which format has to be used.

Change-Id: I9546993987b873e8ae921664238b234608e37bba
Related: OS#5760