PCUIF_Components: ensure clean IMSI string
When receiving an IMSI from PCUIF (see type record PCUIF_pch), it is
represented as a null terminated string. The field is set to be 17
characters wide with a pdding of zeros at the end (as it ought to be
for a null terminated string). Unfortunately TTCN3 will not chop off
the trailing zeros, and also include them when the string length is
determined using lengthof(). This means we must take care of this
ourselves.

Let's use a regular expression to make sure any non numerical digits
are trimmed off before passing the IMSI string on to higher layers.

Related: OS#5927
Change-Id: I7bfea59a306e75211856e4e80985ebf000c42224